+++ ai/model/src/plan_model.py
@@ -16,19 +16,16 @@ from __future__ import annotations
 import json
 import logging
 import os
-from datetime import date
+from datetime import date, datetime, timedelta
 from pathlib import Path
 from typing import Any, Dict, List, Optional
-import time
 
 from .promts.plan_promt import JSON_SCHEMA_EXAMPLE, system_plan_prompt
+from .utils.helper_function import extract_image_url
 
 from dotenv import load_dotenv
-from langchain.agents import Tool, initialize_agent
-from langchain.chains import LLMChain
 from langchain_core.output_parsers import JsonOutputParser
 from langchain.prompts import PromptTemplate
-from langchain_openai import OpenAI
 from langchain.chat_models import ChatOpenAI
 
 from utils.utils import save_data_to_json
@@ -47,25 +44,6 @@ FORMAT_INSTRUCTIONS = (
     f"{json.dumps(JSON_SCHEMA_EXAMPLE, ensure_ascii=False)}"
 )
 
-json_parser = JsonOutputParser()
-
-# ---------------------------------------------------------------------------
-# 🛠️  Optional tools stubs – plug real API later
-# ---------------------------------------------------------------------------
-
-def dummy_weather_tool(query: str) -> str:  # pragma: no cover
-    """Placeholder weather tool."""
-    return "{\"temp\":30, \"condition\":\"sunny\"}"
-
-TOOLS: List[Tool] = [
-    Tool(name="weather", func=dummy_weather_tool, description="Lấy dữ liệu thời tiết"),
-]
-
-# ---------------------------------------------------------------------------
-# 🚂  Planner class
-# ---------------------------------------------------------------------------
-
-
 class PlanModel:
     def __init__(self, temperature: float = 0.7):
         self.llm = ChatOpenAI(
@@ -74,11 +52,9 @@ class PlanModel:
             model="gpt-4",  
             max_tokens=4000  
         )
-        self.parser = json_parser  # langchain JSON parser
+        self.parser = JsonOutputParser()  
 
-    # ---------------------------------------------------------------------
-    # 🔑 Prompt builder
-    # ---------------------------------------------------------------------
+
     def _build_prompt(self) -> PromptTemplate:
         template = (
             "You are an expert Vietnamese travel planner. Using the user data, "
@@ -88,12 +64,150 @@ class PlanModel:
             "Do NOT add title, notes, or explanations.\n\n"
             "User context (JSON): {user_json}\n\n" + FORMAT_INSTRUCTIONS
         )
-        # Specify the input variables explicitly
         return PromptTemplate(template=template, input_variables=["user_json"])
 
-    # ------------------------------------------------------------------
-    # 🧩  Plain chain (no external tools)
-    # ------------------------------------------------------------------
+    def _build_day_prompt(self, day_num: int, current_date_str: str, merged_data: Dict[str, Any]) -> str:
+        """Build the prompt for generating a specific day's plan.
+        
+        Args:
+            day_num: The day number (0-based)
+            current_date_str: The current date in YYYY-MM-DD format
+            merged_data: The merged input data containing accommodations, places, restaurants
+            
+        Returns:
+            str: The complete prompt for generating the day's plan
+        """
+        self.day_title = f"Ngày {day_num+1}: "
+        if day_num == 0:
+            self.day_title += "Khám phá biển"
+        elif day_num == 1:
+            self.day_title += "Khám phá núi"
+        elif day_num == 2:
+            self.day_title += "Khám phá văn hóa"
+        else:
+            self.day_title += "Khám phá địa phương"
+        
+        day_prompt = f"""
+        Tạo chi tiết cho ngày {day_num+1} (ngày {current_date_str}) của lịch trình du lịch {merged_data.get("destination")}.
+        Tạo 3 segments (morning, afternoon, evening) với các hoạt động phù hợp.
+        
+        CHÚ Ý QUAN TRỌNG: 
+        1. CHỈ TRẢ VỀ JSON THUẦN TÚY! KHÔNG THÊM BẤT KỲ VĂN BẢN NÀO TRƯỚC HOẶC SAU JSON!
+        2. PHẢN HỒI CỦA BẠN PHẢI BẮT ĐẦU BẰNG DẤU "{{" VÀ KẾT THÚC BẰNG DẤU "}}" - KHÔNG CÓ GÌ KHÁC!
+        3. MÔ TẢ PHẢI NGẮN GỌN (<100 ký tự) để tránh vượt quá giới hạn token
+        4. KHÔNG sử dụng mô tả dài, CHỈ 1-2 câu ngắn gọn
+        5. JSON KHÔNG ĐƯỢC CẮT NGẮN GIỮA CHỪNG - KIỂM TRA KỸ TẤT CẢ DẤU NGOẶC ĐỀU ĐƯỢC ĐÓNG!
+        
+        Thông tin chuyến đi:
+        Điểm đến: {merged_data.get("destination")}
+        Khách sạn: {[(acc.get("name", ""), acc.get("accommodation_id", "")) for acc in merged_data.get("accommodations", [])]}
+        Địa điểm: {[(place.get("name", ""), place.get("place_id", "")) for place in merged_data.get("places", [])]}
+        Nhà hàng: {[(rest.get("name", ""), rest.get("restaurant_id", "")) for rest in merged_data.get("restaurants", [])]}
+        
+        Cấu trúc JSON cần tuân thủ:
+        {{
+            "date": "{current_date_str}",
+            "day_title": "Ngày {day_num+1}: [Tiêu đề ngắn gọn]",
+            "segments": [
+                {{
+                    "time_of_day": "morning",
+                    "activities": [
+                        {{
+                            "id": "[ID CHÍNH XÁC từ dữ liệu]",
+                            "type": "accommodation",
+                            "name": "Tên hoạt động",
+                            "start_time": "08:00",
+                            "end_time": "10:00",
+                            "description": "Bạn sẽ được...",
+                            // Các trường khác tùy loại hoạt động
+                        }}
+                    ]
+                }},
+                // Tương tự cho afternoon và evening
+            ]
+        }}
+        
+        Hướng dẫn quan trọng:
+        - Tiêu đề ngày PHẢI NGẮN GỌN và sáng tạo (ví dụ: "Khám phá Hà Nội cổ kính")
+        - GIẢM ĐỘ DÀI mô tả, chỉ cần 1-2 câu ngắn với phong cách hướng dẫn viên
+        - Luôn sử dụng đúng ID từ dữ liệu đầu vào
+        - Mỗi segment có 1-2 hoạt động (KHÔNG cần 3 hoạt động/segment để giảm kích thước JSON)
+        - XÓA tất cả chú thích, hướng dẫn trong JSON cuối cùng
+        - Description phải ngắn gọn sáng tạo và có thể chèn thêm icon. 
+        """
+        
+        day_prompt += f"""
+        
+        Đây là ví dụ chuẩn về JSON ngắn gọn cần tạo (NHƯNG PHẢI THAY BẰNG DỮ LIỆU THỰC TẾ):
+        
+        {{
+            "date": "{current_date_str}",
+            "day_title": "Ngày {day_num+1}: Khám phá Hà Nội",
+            "segments": [
+                {{
+                    "time_of_day": "morning",
+                    "activities": [
+                        {{
+                            "id": "hotel_123",
+                            "type": "accommodation",
+                            "name": "Khách sạn ABC",
+                            "start_time": "08:00", 
+                            "end_time": "10:00",
+                            "description": "Bạn sẽ được tận hưởng không gian nghỉ dưỡng thoải mái.",
+                            "location": "Hà Nội",
+                            "rating": 4.5,
+                            "price": 850000,
+                            "image_url": "",
+                            "url": ""
+                        }}
+                    ]
+                }},
+                {{
+                    "time_of_day": "afternoon",
+                    "activities": [
+                        {{
+                            "id": "place_456",
+                            "type": "place",
+                            "name": "Địa điểm XYZ",
+                            "start_time": "13:00",
+                            "end_time": "15:00",
+                            "description": "Hãy khám phá nét văn hóa đặc sắc tại địa điểm này.",
+                            "address": "Hà Nội",
+                            "categories": "sightseeing",
+                            "rating": 4.5,
+                            "price": 50000,
+                            "image_url": "",
+                            "url": ""
+                        }}
+                    ]
+                }},
+                {{
+                    "time_of_day": "evening",
+                    "activities": [
+                        {{
+                            "id": "restaurant_789",
+                            "type": "restaurant",
+                            "name": "Nhà hàng XYZ",
+                            "start_time": "18:00",
+                            "end_time": "20:00",
+                            "description": "Thưởng thức ẩm thực đặc sắc tại nhà hàng nổi tiếng.",
+                            "address": "Hà Nội",
+                            "cuisines": "Đặc sản địa phương",
+                            "rating": 4.5,
+                            "phone": "",
+                            "image_url": "",
+                            "url": ""
+                        }}
+                    ]
+                }}
+            ]
+        }}
+        
+        NHẮC LẠI: JSON phải ngắn gọn và hoàn chỉnh, không được có chú thích hay bị thiếu dấu ngoặc.
+        """
+        
+        return day_prompt
+
     def generate_plan(self, input_data: Dict[str, Any], **meta: Any) -> Dict[str, Any]:
         """LLM only – returns parsed JSON dict."""
         log.info("Generating plan (no agent)…")
@@ -148,137 +262,7 @@ class PlanModel:
                 current_date = start_date + timedelta(days=day_num)
                 current_date_str = current_date.strftime("%Y-%m-%d")
                 
-                day_title = f"Ngày {day_num+1}: "
-                if day_num == 0:
-                    day_title += "Khám phá biển"
-                elif day_num == 1:
-                    day_title += "Khám phá núi"
-                elif day_num == 2:
-                    day_title += "Khám phá văn hóa"
-                else:
-                    day_title += "Khám phá địa phương"
-                
-                # Create prompt for this specific day with simplified structure
-                day_prompt = f"""
-                Tạo chi tiết cho ngày {day_num+1} (ngày {current_date_str}) của lịch trình du lịch {merged_data.get("destination")}.
-                Tạo 3 segments (morning, afternoon, evening) với các hoạt động phù hợp.
-                
-                CHÚ Ý QUAN TRỌNG: 
-                1. CHỈ TRẢ VỀ JSON THUẦN TÚY! KHÔNG THÊM BẤT KỲ VĂN BẢN NÀO TRƯỚC HOẶC SAU JSON!
-                2. PHẢN HỒI CỦA BẠN PHẢI BẮT ĐẦU BẰNG DẤU "{{" VÀ KẾT THÚC BẰNG DẤU "}}" - KHÔNG CÓ GÌ KHÁC!
-                3. MÔ TẢ PHẢI NGẮN GỌN (<100 ký tự) để tránh vượt quá giới hạn token
-                4. KHÔNG sử dụng mô tả dài, CHỈ 1-2 câu ngắn gọn
-                5. JSON KHÔNG ĐƯỢC CẮT NGẮN GIỮA CHỪNG - KIỂM TRA KỸ TẤT CẢ DẤU NGOẶC ĐỀU ĐƯỢC ĐÓNG!
-                
-                Thông tin chuyến đi:
-                Điểm đến: {merged_data.get("destination")}
-                Khách sạn: {[(acc.get("name", ""), acc.get("accommodation_id", "")) for acc in merged_data.get("accommodations", [])]}
-                Địa điểm: {[(place.get("name", ""), place.get("place_id", "")) for place in merged_data.get("places", [])]}
-                Nhà hàng: {[(rest.get("name", ""), rest.get("restaurant_id", "")) for rest in merged_data.get("restaurants", [])]}
-                
-                Cấu trúc JSON cần tuân thủ:
-                {{
-                    "date": "{current_date_str}",
-                    "day_title": "Ngày {day_num+1}: [Tiêu đề ngắn gọn]",
-                    "segments": [
-                        {{
-                            "time_of_day": "morning",
-                            "activities": [
-                                {{
-                                    "id": "[ID CHÍNH XÁC từ dữ liệu]",
-                                    "type": "accommodation",
-                                    "name": "Tên hoạt động",
-                                    "start_time": "08:00",
-                                    "end_time": "10:00",
-                                    "description": "Bạn sẽ được...",
-                                    // Các trường khác tùy loại hoạt động
-                                }}
-                            ]
-                        }},
-                        // Tương tự cho afternoon và evening
-                    ]
-                }}
-                
-                Hướng dẫn quan trọng:
-                - Tiêu đề ngày PHẢI NGẮN GỌN và sáng tạo (ví dụ: "Khám phá Hà Nội cổ kính")
-                - GIẢM ĐỘ DÀI mô tả, chỉ cần 1-2 câu ngắn với phong cách hướng dẫn viên
-                - Luôn sử dụng đúng ID từ dữ liệu đầu vào
-                - Mỗi segment có 1-2 hoạt động (KHÔNG cần 3 hoạt động/segment để giảm kích thước JSON)
-                - XÓA tất cả chú thích, hướng dẫn trong JSON cuối cùng
-                - Description phải ngắn gọn sáng tạo và có thể chèn thêm icon. 
-                """
-                
-                # Điều chỉnh prompt dựa vào day_num
-                # Thêm ví dụ JSON hoàn chỉnh tối giản để model dễ dàng tham khảo
-                day_prompt += f"""
-                
-                Đây là ví dụ chuẩn về JSON ngắn gọn cần tạo (NHƯNG PHẢI THAY BẰNG DỮ LIỆU THỰC TẾ):
-                
-                {{
-                    "date": "{current_date_str}",
-                    "day_title": "Ngày {day_num+1}: Khám phá Hà Nội",
-                    "segments": [
-                        {{
-                            "time_of_day": "morning",
-                            "activities": [
-                                {{
-                                    "id": "hotel_123",
-                                    "type": "accommodation",
-                                    "name": "Khách sạn ABC",
-                                    "start_time": "08:00", 
-                                    "end_time": "10:00",
-                                    "description": "Bạn sẽ được tận hưởng không gian nghỉ dưỡng thoải mái.",
-                                    "location": "Hà Nội",
-                                    "rating": 4.5,
-                                    "price": 850000,
-                                    "image_url": "",
-                                    "url": ""
-                                }}
-                            ]
-                        }},
-                        {{
-                            "time_of_day": "afternoon",
-                            "activities": [
-                                {{
-                                    "id": "place_456",
-                                    "type": "place",
-                                    "name": "Địa điểm XYZ",
-                                    "start_time": "13:00",
-                                    "end_time": "15:00",
-                                    "description": "Hãy khám phá nét văn hóa đặc sắc tại địa điểm này.",
-                                    "address": "Hà Nội",
-                                    "categories": "sightseeing",
-                                    "rating": 4.5,
-                                    "price": 50000,
-                                    "image_url": "",
-                                    "url": ""
-                                }}
-                            ]
-                        }},
-                        {{
-                            "time_of_day": "evening",
-                            "activities": [
-                                {{
-                                    "id": "restaurant_789",
-                                    "type": "restaurant",
-                                    "name": "Nhà hàng XYZ",
-                                    "start_time": "18:00",
-                                    "end_time": "20:00",
-                                    "description": "Thưởng thức ẩm thực đặc sắc tại nhà hàng nổi tiếng.",
-                                    "address": "Hà Nội",
-                                    "cuisines": "Đặc sản địa phương",
-                                    "rating": 4.5,
-                                    "phone": "",
-                                    "image_url": "",
-                                    "url": ""
-                                }}
-                            ]
-                        }}
-                    ]
-                }}
-                
-                NHẮC LẠI: JSON phải ngắn gọn và hoàn chỉnh, không được có chú thích hay bị thiếu dấu ngoặc.
-                """
+                day_prompt = self._build_day_prompt(day_num, current_date_str, merged_data)
                 
                 messages = [
                     {"role": "system", "content": system_prompt},
@@ -288,10 +272,8 @@ class PlanModel:
                 try:
                     day_response = self.llm.invoke(messages)
                     
-                    # Lấy nội dung từ day_response
                     day_response_content = day_response.content if hasattr(day_response, 'content') else day_response
                     
-                    # Chuẩn hóa kết quả trước khi parse
                     day_response_content = self._cleanup_llm_response(day_response_content)
                     
                     try:
@@ -346,57 +328,25 @@ class PlanModel:
                         
                         existing_segments = {segment.get("time_of_day"): segment for segment in day_data["segments"]}
                         
-                        # Check for required segments and add if missing
                         for required_segment in ["morning", "afternoon", "evening"]:
                             if required_segment not in existing_segments:
-                                # Define a helper function to extract image URL from various formats
-                                def extract_image_url(item):
-                                    # Check if there's a direct image_url string
-                                    if isinstance(item.get("image_url"), str) and item.get("image_url"):
-                                        return item.get("image_url")
-                                    
-                                    # Check alternative direct image fields
-                                    if isinstance(item.get("imageUrl"), str) and item.get("imageUrl"):
-                                        return item.get("imageUrl")
-                                    if isinstance(item.get("image"), str) and item.get("image"):
-                                        return item.get("image")
-                                    
-                                    # Check for array-based image fields
-                                    for field in ["image_url", "imageUrl", "image", "images"]:
-                                        # If the field exists and is a list/array
-                                        if isinstance(item.get(field), list) and len(item.get(field)) > 0:
-                                            first_image = item.get(field)[0]
-                                            # If the item is a string, use it directly
-                                            if isinstance(first_image, str):
-                                                return first_image
-                                            # If the item is a dict with a url field
-                                            elif isinstance(first_image, dict) and first_image.get("url"):
-                                                return first_image.get("url")
-                                    
-                                    # Nothing found
-                                    return ""
+                                
 
                                 if required_segment not in existing_segments:
-                                    # Create a basic segment with default activity
                                     default_activity = {}
                                     if required_segment == "morning" and merged_data.get("accommodations"):
-                                        # Get accommodation ID first
                                         accommodation_id = merged_data["accommodations"][0].get("accommodation_id", merged_data["accommodations"][0].get("id", f"hotel_morning_day{day_num+1}"))
                                         
-                                        # Print all accommodation IDs for debugging
                                         log.info(f"Available accommodation IDs: {[a.get('accommodation_id', a.get('id', 'unknown')) for a in merged_data.get('accommodations', [])]}")
                                         log.info(f"Looking for accommodation ID: {accommodation_id}")
                                         
-                                        # Find the matching accommodation to get complete data
                                         matching_accommodation = None
                                         for acc in merged_data.get("accommodations", []):
-                                            # Check for exact matches first
                                             if acc.get("accommodation_id") == accommodation_id or acc.get("id") == accommodation_id:
                                                 matching_accommodation = acc
                                                 log.info(f"Found exact match for accommodation ID: {accommodation_id}")
                                                 break
                                             
-                                            # If the accommodation_id contains the ID (e.g., "hotel_123" contains "123")
                                             elif accommodation_id and acc.get("accommodation_id") and accommodation_id in acc.get("accommodation_id"):
                                                 matching_accommodation = acc
                                                 log.info(f"Found partial match: {accommodation_id} in {acc.get('accommodation_id')}")
@@ -419,8 +369,6 @@ class PlanModel:
                                         
                                         original_description = matching_accommodation.get("description", "")
                                         if original_description:
-                                            # Sử dụng nội dung từ original_description nhưng để model tự viết lại
-                                            # với giọng văn hướng dẫn viên du lịch
                                             key_points = original_description[:200] if len(original_description) > 200 else original_description
                                             description = f"Tại khách sạn tuyệt vời này, bạn sẽ được tận hưởng {key_points.split('.')[0].lower() if '.' in key_points else key_points.lower()}. Hãy nghỉ ngơi và chuẩn bị cho những trải nghiệm tuyệt vời tiếp theo!"
                                         else:
@@ -653,13 +601,13 @@ class PlanModel:
                         # Create a basic day structure
                         basic_day = {
                             "date": current_date_str,
-                            "day_title": day_title,
+                            "day_title": self.day_title,
                             "segments": [
                                 {"time_of_day": "morning", "activities": []},
                                 {"time_of_day": "afternoon", "activities": []},
                                 {"time_of_day": "evening", "activities": []}
                             ]
-                        }
+                        }   
                         # Bổ sung các hoạt động mặc định cho ngày này
                         basic_day = self._populate_default_activities(basic_day, day_num, merged_data)
                         final_plan["plan_by_day"].append(basic_day)
@@ -668,26 +616,22 @@ class PlanModel:
                     # Create a basic day structure
                     basic_day = {
                         "date": current_date_str,
-                        "day_title": day_title,
+                        "day_title": self.day_title,
                         "segments": [
                             {"time_of_day": "morning", "activities": []},
                             {"time_of_day": "afternoon", "activities": []},
                             {"time_of_day": "evening", "activities": []}
                         ]
                     }
-                    # Bổ sung các hoạt động mặc định cho ngày này
                     basic_day = self._populate_default_activities(basic_day, day_num, merged_data)
                     final_plan["plan_by_day"].append(basic_day)
             
-            # Đảm bảo tất cả các ngày đều có dữ liệu đầy đủ
             for idx, day in enumerate(final_plan["plan_by_day"]):
-                # Kiểm tra các segment
                 empty_segments = []
                 for segment in day.get("segments", []):
                     if not segment.get("activities"):
                         empty_segments.append(segment["time_of_day"])
-                
-                # Nếu có segment trống, làm đầy chúng
+            
                 if empty_segments:
                     log.info(f"Ngày {idx+1} có {len(empty_segments)} segment trống, đang bổ sung hoạt động mặc định")
                     day = self._populate_default_activities(day, idx, merged_data)
@@ -700,8 +644,7 @@ class PlanModel:
             log.debug(f"Input data: {input_data}")
             log.debug(f"Meta data: {meta}")
             
-            # Return a basic structure in case of error
-            from datetime import datetime
+
             return {
                 "error": str(e),
                 "trip_name": input_data.get("trip_name", meta.get("trip_name", "Trip Plan")),
@@ -718,12 +661,10 @@ class PlanModel:
         if current_hour is None:
             current_hour = datetime.now().hour
         
-        # Định nghĩa khung giờ cho các segment
         morning_hours = range(5, 12)   # 5:00 - 11:59
         afternoon_hours = range(12, 18) # 12:00 - 17:59
         evening_hours = range(18, 23)  # 18:00 - 22:59
         
-        # Xác định segment hiện tại và các segment còn lại
         current_segment = None
         if current_hour in morning_hours:
             current_segment = "morning"
@@ -750,7 +691,6 @@ class PlanModel:
         }
         
         adjusted_times = {}
-        
         if current_segment == "morning":
             start_hour = max(8, current_hour + 1)  
             adjusted_times["morning"] = [
@@ -767,25 +707,22 @@ class PlanModel:
                 {"start_time": f"{start_hour+2:02d}:00", "end_time": f"{start_hour+3:02d}:00"}
             ]
             adjusted_times["evening"] = segment_times["evening"]
-            adjusted_times["morning"] = None  # Đánh dấu buổi sáng không có kế hoạch
+            adjusted_times["morning"] = None  
             
         elif current_segment == "evening":
-            # Bỏ qua buổi sáng và chiều, chỉ lên kế hoạch cho buổi tối
-            start_hour = max(18, current_hour + 1)  # Bắt đầu ít nhất 1 giờ sau giờ hiện tại
+            start_hour = max(18, current_hour + 1) 
             adjusted_times["evening"] = [
                 {"start_time": f"{start_hour:02d}:00", "end_time": f"{start_hour+1:02d}:30"},
                 {"start_time": f"{start_hour+2:02d}:00", "end_time": f"{min(start_hour+3, 23):02d}:00"}
             ]
-            adjusted_times["morning"] = None  # Đánh dấu buổi sáng không có kế hoạch
-            adjusted_times["afternoon"] = None  # Đánh dấu buổi chiều không có kế hoạch
+            adjusted_times["morning"] = None  
+            adjusted_times["afternoon"] = None  
             
-        else:  # next_day - qua nửa đêm, lên kế hoạch cho ngày mai
-            # Giữ nguyên tất cả thời gian mặc định
+        else:  
             adjusted_times = segment_times
         
         return adjusted_times, current_segment
 
-    # Thêm hàm xử lý cleanup kết quả từ OpenAI trước khi parse JSON
     def _cleanup_llm_response(self, response_text):
         """
         Làm sạch kết quả từ OpenAI API, đảm bảo chỉ trả về JSON hợp lệ.
@@ -813,19 +750,17 @@ class PlanModel:
         
         for i, char in enumerate(cleaned_text):
             if char == '{':
-                if not stack:  # If this is the first opening brace
+                if not stack:  
                     start_idx = i
                 stack.append('{')
             elif char == '}':
                 if stack and stack[-1] == '{':
                     stack.pop()
-                    if not stack:  # If we've closed all braces
+                    if not stack:  
                         potential_jsons.append(cleaned_text[start_idx:i+1])
         
-        # Sort by length (longest first) as it's likely to be more complete
         for json_str in sorted(potential_jsons, key=len, reverse=True):
             try:
-                # Try to parse the JSON directly
                 json_obj = json.loads(json_str)
                 log.info(f"Found valid JSON of length {len(json_str)}")
                 return json_str
@@ -854,7 +789,6 @@ class PlanModel:
             title_match = re.search(r'"day_title"\s*:\s*"([^"]+)"', json_fragment)
             
             if date_match or title_match:
-                # Start building a valid JSON structure
                 partial_json = {
                     "date": date_match.group(1) if date_match else "",
                     "day_title": title_match.group(1) if title_match else "",
@@ -932,17 +866,13 @@ class PlanModel:
                     log.info("Extracted JSON from first '{' to last '}', length: " + str(len(json_candidate)))
                     return json_candidate
                 except json.JSONDecodeError:
-                    # Try fixing common issues
-                    # Balance braces
                     open_count = json_candidate.count('{')
                     close_count = json_candidate.count('}')
                     if open_count > close_count:
                         json_candidate += '}' * (open_count - close_count)
                     
-                    # Fix trailing commas
                     json_candidate = re.sub(r',\s*([}\]])', r'\1', json_candidate)
                     
-                    # Try parsing again
                     try:
                         json.loads(json_candidate)
                         log.info(f"Fixed extracted JSON, length: {len(json_candidate)}")
@@ -956,9 +886,7 @@ class PlanModel:
         if match:
             try:
                 json_str = match.group(1)
-                # Remove any trailing or leading non-JSON text
                 json_str = re.sub(r'[^{}[\],:"0-9a-zA-Z_\-.\s]+', ' ', json_str)
-                # Try to parse
                 json.loads(json_str)
                 log.info(f"Found valid JSON with regex approach, length {len(json_str)}")
                 return json_str
@@ -1010,46 +938,22 @@ class PlanModel:
         
         existing_segments = {segment.get("time_of_day"): segment for segment in day_data.get("segments", [])}
         
-        def extract_image_url(item):
-            if isinstance(item.get("image_url"), str) and item.get("image_url"):
-                return item.get("image_url")
-            
-            if isinstance(item.get("imageUrl"), str) and item.get("imageUrl"):
-                return item.get("imageUrl")
-            
-            if isinstance(item.get("image"), str) and item.get("image"):
-                return item.get("image")
-            
-            for field in ["image_url", "imageUrl", "image", "images"]:
-                if isinstance(item.get(field), list) and len(item.get(field)) > 0:
-                    first_image = item.get(field)[0]
-                    if isinstance(first_image, str):
-                        return first_image
-                    elif isinstance(first_image, dict) and first_image.get("url"):
-                        return first_image.get("url")
-                    
-            return ""
         
-        # Thêm mặc định cho cả 3 segments
         for segment_type in ["morning", "afternoon", "evening"]:
             segment = existing_segments.get(segment_type)
             
-            # Nếu segment không tồn tại, tạo mới
             if not segment:
                 segment = {"time_of_day": segment_type, "activities": []}
                 day_data["segments"].append(segment)
                 existing_segments[segment_type] = segment
             
-            # Đảm bảo mỗi segment đều có hoạt động
             if not segment.get("activities"):
                 activities = []
                 
-                # Morning: thêm khách sạn
                 if segment_type == "morning" and merged_data.get("accommodations"):
                     accommodation = merged_data["accommodations"][0]
                     accommodation_id = accommodation.get("accommodation_id", accommodation.get("id", f"hotel_morning_day{day_num+1}"))
                     
-                    # Tìm accommodation phù hợp
                     matching_accommodation = None
                     for acc in merged_data.get("accommodations", []):
                         if (acc.get("accommodation_id") == accommodation_id or 
@@ -1084,7 +988,6 @@ class PlanModel:
                         "url": matching_accommodation.get("url", matching_accommodation.get("link", "")) if matching_accommodation else ""
                     })
                 
-                # Afternoon: thêm địa điểm tham quan
                 elif segment_type == "afternoon" and merged_data.get("places"):
                     place_index = min(day_num, len(merged_data["places"])-1) if merged_data["places"] else 0
                     if place_index >= 0 and merged_data["places"]:
@@ -1112,7 +1015,6 @@ class PlanModel:
                             "url": place.get("url", place.get("link", ""))
                         })
                 
-                # Evening: thêm nhà hàng
                 elif segment_type == "evening" and merged_data.get("restaurants"):
                     rest_index = min(day_num, len(merged_data["restaurants"])-1) if merged_data["restaurants"] else 0
                     if rest_index >= 0 and merged_data["restaurants"]:
@@ -1152,9 +1054,6 @@ class PlanModel:
         return day_data
 
 
-# ---------------------------------------------------------------------------
-# 🧪 CLI quick test (python travel_planner.py)
-# ---------------------------------------------------------------------------
 if __name__ == "__main__":
     sample_input = {
         "accommodations": [{"id": "hotel42", "name": "Sala", "price": 850000}],
