+++ ai/model/src/vector_database/script/prepare_embedding_data/setup_hotel_data.sh
